refactor: GameModeStats pulled out of Stats fix: stats save on GameOver
