+++ src/main/java/org/example/entities/player/PlayerDbService.java
@@ -4,11 +4,12 @@ import org.example.entities.user.User;
 
 public class PlayerDbService {
 
-  public Player toPlayer(User user, String connectionId, Boolean isWhite) {
+  public Player toPlayer(User user, int rating, String connectionId, Boolean isWhite) {
     return Player.builder()
         .playerId(user.getId())
         .username(user.getUsername())
         .remainingTime(1)
+        .rating(rating)
         .connectionId(connectionId)
         .isWhite(isWhite)
         .build();
