+++ src/main/java/org/example/handlers/websocket/makeMove/MakeMoveService.java
@@ -138,9 +138,11 @@ public class MakeMoveService {
   }
 
   /**
-   * if the game arg is in a state of checkmate, this function will call the GameOverService to handle it
+   * if the game arg is in a state of checkmate, this function will call the GameOverService to
+   * handle it
+   *
    * @return true if it is a checkmate, false otherwise
-   * */
+   */
   public boolean handleCheckmate(Game game, SocketMessenger messenger) throws InternalServerError {
     Board board = Board.fromFEN(game.getGameStateAsFen());
 
@@ -170,8 +172,9 @@ public class MakeMoveService {
 
   /**
    * if the game arg is in a state of draw, this function will call the GameOverService to handle it
+   *
    * @return true if game is drawn, false otherwise
-   * */
+   */
   public boolean handleDraw(Game game, SocketMessenger messenger) throws InternalServerError {
     Board board = Board.fromFEN(game.getGameStateAsFen());
